+++ packages/angular/src/index.ts
@@ -1,41 +1,8 @@
 export { create } from "./create";
 
-export type {
-  SpooshInstanceShape,
-  EnabledOption,
-  AngularOptionsMap,
-} from "./types/shared";
-export type {
-  ExtractMethodData,
-  ExtractMethodError,
-  ExtractMethodOptions,
-  ExtractMethodQuery,
-  ExtractMethodBody,
-  ExtractResponseQuery,
-  ExtractResponseBody,
-  ExtractResponseParamNames,
-} from "./types/extraction";
-export type {
-  BaseReadOptions,
-  BaseReadResult,
-  ReadApiClient,
-  TriggerOptions,
-  ResponseInputFields,
-} from "./injectRead/types";
-export type {
-  BaseWriteResult,
-  WriteApiClient,
-  WriteResponseInputFields,
-} from "./injectWrite/types";
-export type {
-  BasePagesOptions,
-  BasePagesResult,
-  PagesApiClient,
-  PagesTriggerOptions,
-} from "./injectPages/types";
-export type {
-  BaseQueueResult,
-  QueueApiClient,
-  QueueTriggerInput,
-  InjectQueueOptions,
-} from "./injectQueue/types";
+export type * from "./types/shared";
+export type * from "./types/extraction";
+export type * from "./injectRead/types";
+export type * from "./injectWrite/types";
+export type * from "./injectPages/types";
+export type * from "./injectQueue/types";